review feedback - Add dom_wakelock_enabled pref (disabled by default) to gate the API      until embedder integration is complete    - Move wakelock.rs and wakelocksentinel.rs into dom/wakelock/ subfolder      following the pattern of other multi-file DOM modules (e.g. gamepad)    - Add Pref=dom_wakelock_enabled to Navigator.wakeLock, WakeLock, and      WakeLockSentinel in WebIDL    - Annotate WakeLock.request() with cx in Bindings.conf; use      reflect_dom_object_with_cx in WakeLock::new()    - Annotate Navigator.wakeLock with cx in Bindings.conf    - Add MPL license header and spec link to WakeLock.webidl    - Add screen-wake-lock to tests/wpt/include.ini and      tests/wpt/meta/screen-wake-lock/__dir__.ini with pref enabled
