+++ components/script/dom/mod.rs
@@ -377,7 +377,7 @@ pub(crate) mod virtualmethods;
 pub(crate) mod visibilitystateentry;
 pub(crate) mod visualviewport;
 pub(crate) mod wakelock;
-pub(crate) mod wakelocksentinel;
+pub(crate) use self::wakelock::wakelocksentinel;
 pub(crate) mod webgl;
 pub(crate) use self::webgl::extensions::ext::*;
 pub(crate) use self::webgl::*;
