+++ components/script/dom/navigator.rs
@@ -615,9 +615,9 @@ impl NavigatorMethods<crate::DomTypeHolder> for Navigator {
     }
 
     /// <https://w3c.github.io/screen-wake-lock/#dom-navigator-wakelock>
-    fn WakeLock(&self) -> DomRoot<WakeLock> {
+    fn WakeLock(&self, cx: &mut js::context::JSContext) -> DomRoot<WakeLock> {
         self.wake_lock
-            .or_init(|| WakeLock::new(&self.global(), CanGc::note()))
+            .or_init(|| WakeLock::new(cx, &self.global(), CanGc::note()))
     }
 }
