+++ components/script_bindings/codegen/Bindings.conf
@@ -683,7 +683,11 @@ DOMInterfaces = {
 'Navigator': {
     'inRealms': ['GetVRDisplays'],
     'canGc': ['Languages', 'SendBeacon', 'UserActivation'],
-    'cx': ['Clipboard']
+    'cx': ['Clipboard', 'WakeLock'],
+},
+
+'WakeLock': {
+    'cx': ['Request'],
 },
 
 'Node': {
