+++ tests/wpt/include.ini
@@ -254,6 +254,8 @@ skip: true
   skip: false
 [secure-contexts]
   skip: false
+[screen-wake-lock]
+  skip: false
 [selection]
   skip: false
 [service-workers]
