+++ tests/wpt/meta/screen-wake-lock/__dir__.ini
@@ -1,0 +1,1 @@
+prefs: [dom_wakelock_enabled: true]
